+++ app/src/main/java/com/arctouch/easybus/data/endpoints/StopsEndpoint.java
@@ -3,7 +3,7 @@ package com.arctouch.easybus.data.endpoints;
 import android.content.Context;
 
 import com.arctouch.easybus.R;
-import com.arctouch.easybus.model.Stop;
+import com.arctouch.easybus.route.stops.Stop;
 import com.arctouch.easybus.route.stops.StopsResult;
 
 import org.springframework.http.HttpEntity;
