+++ app/src/main/java/com/arctouch/easybus/route/Route.java
@@ -1,4 +1,4 @@
-package com.arctouch.easybus.model;
+package com.arctouch.easybus.route;
 
 import java.io.Serializable;
