+++ app/src/main/java/com/arctouch/easybus/route/schedule/ScheduleResult.java
@@ -1,6 +1,5 @@
 package com.arctouch.easybus.route.schedule;
 
 import com.arctouch.easybus.data.ServiceResponse;
-import com.arctouch.easybus.model.ScheduleItem;
 
 public class ScheduleResult extends ServiceResponse<ScheduleItem> {}
