+++ app/src/main/java/com/arctouch/easybus/route/stops/StopsResult.java
@@ -1,6 +1,5 @@
 package com.arctouch.easybus.route.stops;
 
 import com.arctouch.easybus.data.ServiceResponse;
-import com.arctouch.easybus.model.Stop;
 
 public class StopsResult extends ServiceResponse<Stop> {}
