+++ app/src/main/java/com/arctouch/easybus/search/RoutesResult.java
@@ -1,6 +1,6 @@
 package com.arctouch.easybus.search;
 
-import com.arctouch.easybus.model.Route;
+import com.arctouch.easybus.route.Route;
 import com.arctouch.easybus.data.ServiceResponse;
 
 public class RoutesResult extends ServiceResponse<Route> {}
